restructure to include router, pinia and split in components and also some external class-free css in my index.html head
