+++ src/views/HomePage.vue
@@ -1,0 +1,9 @@
+<template>
+    <ButtonsSection/>
+    <ContactsSection/>
+</template>
+  
+<script setup>
+    import ButtonsSection from '@/components/ButtonsSection.vue';
+    import ContactsSection from '@/components/ContactsSection.vue';
+</script>
